[#139] Refactor benchmark so that it is a event service benchmark
