--- benchmarks/event-concept-id-tracker/Cargo.toml
@@ -1,20 +1,0 @@
-[package]
-name = "event-concept-id-tracker-benchmark"
-description = "iceoryx2: [internal] benchmark for the iceoryx2-cal::event::id_tracker"
-categories = { workspace = true }
-edition = { workspace = true }
-homepage = { workspace = true }
-keywords = { workspace = true }
-license = { workspace = true }
-repository = { workspace = true }
-rust-version = { workspace = true }
-version = { workspace = true }
-
-[dependencies]
-iceoryx2-bb-lock-free = { workspace = true }
-iceoryx2-cal = { workspace = true }
-iceoryx2-bb-elementary = { workspace = true }
-iceoryx2-bb-memory = { workspace = true }
-
-pin-init = { workspace = true }
-clap = { workspace = true }
