+++ benchmarks/event/Cargo.toml
@@ -1,0 +1,17 @@
+[package]
+name = "event-benchmark"
+description = "iceoryx2: [internal] benchmark for the iceoryx2 event services"
+categories = { workspace = true }
+edition = { workspace = true }
+homepage = { workspace = true }
+keywords = { workspace = true }
+license = { workspace = true }
+repository = { workspace = true }
+rust-version = { workspace = true }
+version = { workspace = true }
+
+[dependencies]
+iceoryx2 = { workspace = true }
+iceoryx2-bb-log = { workspace = true }
+
+clap = { workspace = true }
